commentage des transormation pour un test

+++ server/src/modules/transform/BasicTransform.cpp
@@ -14,7 +14,7 @@ extern "C"
 	ret = new Translation((double)param->x, (double)param->y, (double)param->z);
 	break;
       case 2:
-	ret = new Rotation(param->x, param->y, param->z);
+	ret = new Rotation((double)param->x, (double)param->y, (double)param->z);
 	break;
       default:
 	break;
